Add Haystack/South debugging and new decorators.
Add the test code for debugging the Haystack/South integration issue,
described at:

* https://github.com/toastdriven/django-haystack/issues/495
* http://south.aeracode.org/ticket/1023 (closed wontfix)

Add the @before and @after decorators to reduce redundancy and risk
of failing to call the original_function, calling with the wrong arguments,
or failing to return the result of the original_function when writing
"simple" monkey patches to execute before or after a function/method
without breaking the flow of execution.

Add the @modify_return_value decorator which works like @after, but
adds the original return value as a parameter of the after_function,
and returns the value returned from the after_function instead.